refactor(ai): decompose ai-install into focused internal modules
